# Use sync.Map and atomics to improve sum performance