@dashpole 2025-09-30
Feel free to review.  I tried to document the code as well as I could.  I plan to present this at the 10/2 SIG meeting so it might be easier to review after that.